refactor: useScroll로 scroll 관리하는 로직 분리
